Moved loop_manager.py to managers/
