+++ managers/__init__.py
@@ -1,3 +1,4 @@
 """Managers."""
 
 from .liveness import LivenessManager  # noqa
+from .loop_manager import LoopManager  # noqa
