+++ managers/loop_manager.py
